requirements.txt
pillow>=10.0.0
pystray>=0.19.0
screeninfo>=0.8.1
pywin32>=306
pyinstaller>=6.0.0